+++ test-suite/TestSrcBreakpointResolve/Program.cs
@@ -10,7 +10,13 @@ namespace TestSrcBreakpointResolve
 {
 class test_constructors
 {
-    int test_field = 5; // bp here! make sure you correct code (test constructor)!
+                         // bp here! make sure you correct code (test constructor)!
+    int test_field1 = 1;
+
+    int test_field2 = 2; // bp here! make sure you correct code (test constructor)!
+
+                         // bp here! make sure you correct code (test constructor)!
+    int test_field3 = 3;
 
     public test_constructors()
     {
@@ -309,15 +315,25 @@ Label.Breakpoint("bp20_2");            numbers.ForEach(delegate(string number) {
                 Context Context = (Context)context;
                 Context.WasBreakpointHit(@"__FILE__:__LINE__", "bp23", false);
 
-                Context.AddManualBreakpointAndAddID(@"__FILE__:__LINE__", "Program.cs", 13); // line number with "int test_field = 5;" code
-                Context.AddManualBreakpointAndAddID(@"__FILE__:__LINE__", "Program.cs", 17); // line number with "int i = 5;" code
+                Context.AddManualBreakpointAndAddID(@"__FILE__:__LINE__", "Program.cs", 13); // line number before "int test_field1 = 1;" code
+                Context.AddManualBreakpointAndAddID(@"__FILE__:__LINE__", "Program.cs", 16); // line number with "int test_field2 = 2;" code
+                Context.AddManualBreakpointAndAddID(@"__FILE__:__LINE__", "Program.cs", 18); // line number before "int test_field3 = 3;" code
+                Context.AddManualBreakpointAndAddID(@"__FILE__:__LINE__", "Program.cs", 23); // line number with "int i = 5;" code
                 Context.SetBreakpointsAndCheckIDs(@"__FILE__:__LINE__");
                 Context.Continue(@"__FILE__:__LINE__");
-                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 13); // line number with "int test_field = 5;" code
+                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 14); // line number with "int test_field1 = 1;" code
                 Context.Continue(@"__FILE__:__LINE__");
-                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 17); // line number with "int i = 5;" code
+                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 16); // line number with "int test_field2 = 2;" code
                 Context.Continue(@"__FILE__:__LINE__");
-                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 13); // line number with "int test_field = 5;" code
+                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 19); // line number with "int test_field3 = 3;" code
+                Context.Continue(@"__FILE__:__LINE__");
+                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 23); // line number with "int i = 5;" code
+                Context.Continue(@"__FILE__:__LINE__");
+                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 14); // line number with "int test_field1 = 1;" code
+                Context.Continue(@"__FILE__:__LINE__");
+                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 16); // line number with "int test_field2 = 2;" code
+                Context.Continue(@"__FILE__:__LINE__");
+                Context.WasManualBreakpointHit(@"__FILE__:__LINE__", "Program.cs", 19); // line number with "int test_field3 = 3;" code
                 Context.Continue(@"__FILE__:__LINE__");
             });